README and repo overhaul for the ROCm 2.0 release
Community feedback has pointed out a number of confusing,
oudated, or missing sections in our ROCm README file. For example,
we do not describe what our ROCm package structure is, or how the
packages and meta-packages fit together. This can make it confusing
for users who do not want to just install rocm-dkms and move on.

Our repo manifest (default.xml) is severely out of date. It is
missing almost all of the current ROCm projects, and it always
pulls from the main development branch. This means we do not have
a pinned manifest that allows you to pull the code from a
particular ROCm reelease. Manifest updated, and the section of the
README discussing it is majorly overhauled (including links for
information/scripts about building the code after downloading it).

Rather than continually grow our version history in the main
README page, this splits off old version information into its own
file.